Fix nightly
